Backend/requirements.txt
fastapi
uvicorn
passlib
mysqlclient
sqlalchemy
bcrypt
psycopg2-binary
alembic
google-genai
PyJWT
numpy
scikit-learn
apscheduler